+++ src/channels/kit/registration.ts
