+++ src/channels/slack/slack.ts
@@ -5,14 +5,14 @@ import type { ChannelModule } from "../../channel.ts";
 import { log } from "../../log.ts";
 import { readBodyCapped } from "../body.ts";
 import { text } from "../respond.ts";
-import { createSeenRing } from "../seen.ts";
-import { createThreadParticipants } from "../thread-participants.ts";
-import { createTaskTracker } from "../tasks.ts";
-import { ensureStateHome } from "../state.ts";
-import { dispatchStop, isStopText } from "../stop-command.ts";
-import { codePointPrefix } from "../text.ts";
-import { createTurnQueue } from "../turn-queue.ts";
-import { createTurnStore } from "../turn-store.ts";
+import { createSeenRing } from "../kit/seen.ts";
+import { createThreadParticipants } from "../kit/thread-participants.ts";
+import { createTaskTracker } from "../kit/tasks.ts";
+import { ensureStateHome } from "../kit/state.ts";
+import { dispatchStop, isStopText } from "../kit/stop-command.ts";
+import { codePointPrefix } from "../kit/text.ts";
+import { createTurnQueue } from "../kit/turn-queue.ts";
+import { createTurnStore } from "../kit/turn-store.ts";
 import { createSlackBotTokenProvider } from "./bot-auth.ts";
 import { collectSlackBufferedFiles, createSlackContextBuffer } from "./context-buffer.ts";
 import { invokeSlackTurn } from "./invoke-turn.ts";
